Fix xss vulnerability PullRequiest from romanobro56
from romanobro56:

Fixes: #557

First, I changed the _sender value to currentPlayer.name to prevent player being able to identify sender as anyone other than themself, which fixes the xss vulnerability and is in of itself a vulnerability.

Then, I checked to make sure that currentPlayer.name cannot be used maliciously in the same way as _sender, which is currently prevented by the validNick function, but since it's not clear the purpose of validNick I decided it's best to sanitize the name separately, in case validNick is changed in the future.